LabExamples/Lab10_Tile_Harita/pyproject.toml
[project]
name = "lab10-tile-harita"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pygame-ce>=2.5.7"]
